// src/types/enums/EUserActionTypes.ts
export enum EUserActionTypes {
  changeEmail = 'changeEmail',
  setDateOfBirth = 'setDateOfBirth',
  setFirstName = 'setFirstName',
  setLastName = 'setLastName',
  addAddress = 'addAddress',
  changeAddress = 'changeAddress',
  removeAddress = 'removeAddress',
  setDefaultShippingAddress = 'setDefaultShippingAddress',
  addShippingAddressId = 'addShippingAddressId',
  removeShippingAddressId = 'removeShippingAddressId',
  setDefaultBillingAddress = 'setDefaultBillingAddress',
  addBillingAddressId = 'addBillingAddressId',
  removeBillingAddressId = 'removeBillingAddressId'
}